Rename: 시큐리티에 관련된 코드 위치 변경
시큐리티에 관련된 코드들이 너무 흩어져있어 한 패키지로 관리하기 위해 파일들을 이동함

Related to: #18
