+++ src/main/java/pjh5365/linuxserviceweb/domain/auth/handler/CustomAuthenticationSuccessHandler.java
@@ -1,4 +1,4 @@
-package pjh5365.linuxserviceweb.handler;
+package pjh5365.linuxserviceweb.domain.auth.handler;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
